requirements.txt
customtkinter
python-dotenv
keyboard
pyautogui
deepgram-sdk
PyAudio
pyinstaller
Pillow
termcolor